KmToMiles.java

+++ KmToMiles.java
@@ -2,7 +2,7 @@ import java.util.*;
 public class KmToMiles{
     public static void main(String args[]){
         Scanner sc=new Scanner(System.in);
-        int kilometers=sc.nextInt();
+        double kilometers=sc.nextInt();
         System.out.print("Distance in Miles: "+kilometers*0.621371);
     }
 }
